requirements.txt
opencv-python-headless==4.10.0.84
Pillow==10.4.0
pillow-heif==0.18.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
numpy==1.26.4
